Split bakers and unpackers for Rhino
