+++ Connectors/Rhino/Speckle.Connectors.RhinoShared/HostApp/RhinoColorUnpacker.cs
@@ -1,20 +1,12 @@
-using Rhino;
+﻿using Rhino;
 using Rhino.DocObjects;
 using Speckle.Connectors.Rhino.Extensions;
 using Speckle.Sdk.Models.Proxies;
 
 namespace Speckle.Connectors.Rhino.HostApp;
 
-/// <summary>
-/// Utility class managing colors on objects and layers. Expects to be a scoped dependency per send or receive operation.
-/// </summary>
-public class RhinoColorManager
+public class RhinoColorUnpacker
 {
-  /// <summary>
-  /// For receive operations
-  /// </summary>
-  public Dictionary<string, (Color, ObjectColorSource)> ObjectColorsIdMap { get; } = new();
-
   /// <summary>
   /// For send operations
   /// </summary>
@@ -113,40 +105,5 @@ public class RhinoColorManager
     }
 
     return ColorProxies.Values.ToList();
-  }
-
-  /// <summary>
-  /// Parse Color Proxies and stores in ObjectColorsIdMap the relationship between object ids and colors
-  /// </summary>
-  /// <param name="colorProxies"></param>
-  public void ParseColors(List<ColorProxy> colorProxies)
-  {
-    foreach (ColorProxy colorProxy in colorProxies)
-    {
-      ObjectColorSource source = ObjectColorSource.ColorFromObject;
-      if (colorProxy["source"] is string proxySource)
-      {
-        switch (proxySource)
-        {
-          case "layer":
-            continue; // skip any colors with source = layer, since object color default source is by layer
-          case "block":
-            source = ObjectColorSource.ColorFromParent;
-            break;
-          case "material":
-            source = ObjectColorSource.ColorFromMaterial;
-            break;
-        }
-      }
-
-      foreach (string objectId in colorProxy.objects)
-      {
-        Color convertedColor = Color.FromArgb(colorProxy.value);
-        if (!ObjectColorsIdMap.TryGetValue(objectId, out (Color, ObjectColorSource) _))
-        {
-          ObjectColorsIdMap.Add(objectId, (convertedColor, source));
-        }
-      }
-    }
   }
 }
