+++ Connectors/Rhino/Speckle.Connectors.RhinoShared/Speckle.Connectors.RhinoShared.projitems
@@ -28,14 +28,18 @@
     <Compile Include="$(MSBuildThisFileDirectory)Extensions\RhinoUnitsExtension.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Filters\RhinoEverythingFilter.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Filters\RhinoSelectionFilter.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)HostApp\RhinoColorBaker.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)HostApp\RhinoColorUnpacker.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)HostApp\RhinoContext.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)HostApp\RhinoDocumentStore.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)HostApp\RhinoGroupManager.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)HostApp\RhinoGroupBaker.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)HostApp\RhinoGroupUnpacker.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)HostApp\RhinoIdleManager.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)HostApp\RhinoInstanceObjectsManager.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)HostApp\RhinoInstanceBaker.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)HostApp\RhinoInstanceUnpacker.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)HostApp\RhinoLayerManager.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)HostApp\RhinoColorManager.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)HostApp\RhinoMaterialManager.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)HostApp\RhinoMaterialBaker.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)HostApp\RhinoMaterialUnpacker.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)HostApp\SpeckleRhinoPanelHost.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Operations\Receive\DisableRedrawScope.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Operations\Receive\RhinoHostObjectBuilder.cs" />
