+++ DUI3/Speckle.Connectors.DUI/ContainerRegistration.cs
@@ -14,6 +14,7 @@ public static class ContainerRegistration
   public static void AddDUI(this SpeckleContainerBuilder speckleContainerBuilder)
   {
     // send operation and dependencies
+    speckleContainerBuilder.AddSingletonInstance<ISyncToThread, SyncToUIThread>();
     speckleContainerBuilder.AddSingleton<IRootObjectSender, RootObjectSender>();
     speckleContainerBuilder.AddTransient<IBridge, BrowserBridge>(); // POC: Each binding should have it's own bridge instance
     speckleContainerBuilder.AddSingleton(GetJsonSerializerSettings());
