+++ Sdk/Speckle.Connectors.Utils/Operations/Receive/RootObjectUnpacker.cs
@@ -1,5 +1,6 @@
 ﻿using Speckle.Objects.Other;
 using Speckle.Sdk.Models;
+using Speckle.Sdk.Models.Collections;
 using Speckle.Sdk.Models.GraphTraversal;
 using Speckle.Sdk.Models.Instances;
 using Speckle.Sdk.Models.Proxies;
@@ -23,7 +24,7 @@ public class RootObjectUnpacker
   }
 
   public IEnumerable<TraversalContext> GetObjectsToConvert(Base root) =>
-    _traverseFunction.Traverse(root).Where(obj => obj.Current is not Sdk.Models.Collections.Collection);
+    _traverseFunction.Traverse(root).Where(obj => obj.Current is not Collection);
 
   public List<ColorProxy>? TryGetColorProxies(Base root) => TryGetProxies<ColorProxy>(root, COLOR_PROXIES_KEY);
 
@@ -34,6 +35,27 @@ public class RootObjectUnpacker
     TryGetProxies<InstanceDefinitionProxy>(root, INSTANCE_DEFINITION_PROXIES_KEY);
 
   public List<GroupProxy>? TryGetGroupProxies(Base root) => TryGetProxies<GroupProxy>(root, GROUP_PROXIES_KEY);
+
+  public (
+    List<TraversalContext> atomicObjects,
+    List<TraversalContext> instanceComponents
+  ) SplitAtomicObjectsAndInstances(IEnumerable<TraversalContext> objectsToSplit)
+  {
+    List<TraversalContext> atomicObjects = new();
+    List<TraversalContext> instanceComponents = new();
+    foreach (TraversalContext tc in objectsToSplit)
+    {
+      if (tc.Current is IInstanceComponent)
+      {
+        instanceComponents.Add(tc);
+      }
+      else
+      {
+        atomicObjects.Add(tc);
+      }
+    }
+    return (atomicObjects, instanceComponents);
+  }
 
   private List<T>? TryGetProxies<T>(Base root, string key) => (root[key] as List<object>)?.Cast<T>().ToList();
 }
